Вынос getUserFromCookie в отдельный файлик
