+++ src/helpers/user/user.ts
@@ -1,11 +1,3 @@
-import Network from '@network/network';
-import Events from '@eventBus/eventbus';
-import {
-    HAVE_USER,
-    REDIRECT_ERROR,
-    HAVNT_USER,
-} from '@eventBus/constants';
-
 import {UserData} from '@interfaces/userData';
 
 /** Информация о пользователе,
@@ -45,29 +37,11 @@ export default class User {
         };
     }
 
-    getFromCookie(): void {
-        const response = Network.user();
-        response.then((response) => {
-            const code = response.code;
-            const data = response.data;
-            switch (code) {
-            case 200:
-                this.isAuth = true;
-                this.username = data.username;
-                this.id = data.id;
-                this.avatar = Network.getUrlFile(data.avatar);
-                this.email = data.email;
-                Events.trigger(HAVE_USER, this.getData());
-                break;
-            case 401:
-                Events.trigger(HAVNT_USER);
-                break;
-            default:
-                Events.trigger(REDIRECT_ERROR, {url: '/error',
-                    err: 'Что-то страшное произошло c нишим сервером...' + ` Он говорит: ${code}`
-                });
-                break;
-            }
-        });
+    setData(user: UserData): void {
+        this.isAuth = user.isAuth;
+        this.username = user.username;
+        this.email = user.username;
+        this.id = user.id;
+        this.avatar = user.avatar;
     }
 }
